fix participate and toasts

+++ client/src/components/Participate/Participate.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import "./Participate.css";
-import { ToastContainer, toast } from "react-toastify";
+import { toast, ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import { useNavigate, useParams } from "react-router";
 import { useAuth } from "../../contexts/AuthContext";
@@ -36,7 +36,8 @@ function Participate() {
         },
       );
 
-      if (response) {
+      if (response.ok) {
+        setIsParticipated(true);
         toast("Vous participez à cet évènement", {
           type: "success",
         });
@@ -51,14 +52,24 @@ function Participate() {
   };
 
   const deleteParticipation = async () => {
+    if (!auth) {
+      navigate("/login", { state: { islogged: false } });
+      return;
+    }
     try {
       const response = await fetch(
         `${import.meta.env.VITE_API_URL}/api/participate/${userId}/${eventId}`,
         {
           method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${auth.token}`,
+          },
         },
       );
-      if (response) {
+
+      if (response.ok) {
+        setIsParticipated(false);
         toast("Vous ne participez plus à cet évènement", { type: "info" });
       } else {
         throw new Error("Erreur lors de la suppression de la participation");
@@ -74,25 +85,17 @@ function Participate() {
       <button
         className="participate-button"
         type="button"
-        onClick={() => {
-          if (isParticipated) {
-            deleteParticipation();
-            setIsParticipated(true);
+        onClick={async () => {
+          if (!isParticipated) {
+            await addParticipation();
           } else {
-            addParticipation();
-            setIsParticipated(false);
+            await deleteParticipation();
           }
         }}
       >
         {isParticipated ? "Je ne participe plus" : "Je participe"}
       </button>
-
-      <ToastContainer
-        position="top-center"
-        theme="colored"
-        autoClose={2000}
-        limit={2}
-      />
+      <ToastContainer theme="colored" position="top-center" limit={2} />
     </>
   );
 }
